Added test stuff
Added .codeclimate.yml and their badge to the readme. Updated the .gitignore to include a lot more for this type of project. Added a bunch to the package.json, like all of the devDependencies for codecov, eslint, mocha, etc. Added travis badge to readme. Added mocha.opts and a framework for the unit test files.